+++ docs/_site/ipynb2markdown.py
@@ -1,6 +1,6 @@
 # IPython Notebook to Markdown conversion script
 #
-# Sebastian Raschka 2014-2018
+# Sebastian Raschka 2014-2019
 # mlxtend Machine Learning Library Extensions
 #
 # Author: Sebastian Raschka <sebastianraschka.com>