+++ mlxtend/classifier/tests/test_stacking_cv_classifier.py
@@ -1,4 +1,4 @@
-# Sebastian Raschka 2014-2018
+# Sebastian Raschka 2014-2019
 # mlxtend Machine Learning Library Extensions
 # Authors: Sebastian Raschka <sebastianraschka.com>
 #          Reiichiro Nakano <github.com/reiinakano>