+++ mlxtend/image/eyepad_align.py
@@ -1,4 +1,4 @@
-# Sebastian Raschka 2014-2018
+# Sebastian Raschka 2014-2019
 # contributor: Vahid Mirjalili
 # mlxtend Machine Learning Library Extensions
 #
@@ -53,9 +53,9 @@ class EyepadAlign(object):
     eye_distance_ : the distance between left and right eyes
         in the target landmarks.
 
-    target_width_ : the width of the transformed output image.
-
     target_height_ : the height of the transformed output image.
+
+    target_width_ : the width of the transformed output image.
 
     For more usage examples, please see
     http://rasbt.github.io/mlxtend/user_guide/image/EyepadAlign/
@@ -92,7 +92,8 @@ class EyepadAlign(object):
         return self
 
     def fit_directory(self, target_img_dir, target_height,
-                      target_width,  file_extensions='.jpg'):
+                      target_width,  file_extension='.jpg',
+                      pre_check=True):
         """
         Calculates the average landmarks for all face images
         in a directory which will then be set as the target landmark set.
@@ -108,6 +109,15 @@ class EyepadAlign(object):
         target_width : int
             Expected image width of the images in the directory
 
+        file_extension str (default='.jpg'): File extension of the image files.
+
+        pre_check Bool (default=True): Checks that each image has the
+            dimensions specificed via `target_height`
+            and `target_width` on the whole directory first to identify
+            potential issues that are recommended
+            to be fixed before proceeding. Raises a warning for each image if
+            dimensions differ from the ones specified and expected.
+
         Returns
         -------
         self : object
@@ -116,23 +126,45 @@ class EyepadAlign(object):
         self.target_height_ = target_height
         self.target_width_ = target_width
 
-        file_list = [os.path.relpath(os.path.join(dirpath, file),
+        file_list = [os.path.relpath(os.path.join(dirpath, f),
                                      target_img_dir)
                      for (dirpath, dirnames, filenames)
                      in os.walk(target_img_dir)
-                     for file in filenames if file.endswith(file_extensions)]
+                     for f in filenames if f.endswith(file_extension)]
+
+        if not len(file_list):
+            raise ValueError('No images found in %s with extension %s.'
+                             % (target_img_dir, file_extension))
+
+        landmarks_list = []
+
+        if pre_check:
+            if self.verbose >= 1:
+                print('Pre-Checking directory for'
+                      ' consistent image dimensions...')
+                pbar = ProgBar(len(file_list))
+            for f in file_list:
+                img = read_image(filename=f, path=target_img_dir)
+                if self.verbose >= 1:
+                    pbar.update()
+                if (img.shape[0] != self.target_height_
+                        or img.shape[1] != self.target_width_):
+                    warnings.warn('Image %s has '
+                                  'dimensions %d x %d '
+                                  'instead of %d x %d.'
+                                  % (f, img.shape[0],
+                                     img.shape[1],
+                                     self.target_height_,
+                                     self.target_width_))
 
         if self.verbose >= 1:
             print("Fitting the average facial landmarks "
                   "for %d face images " % (len(file_list)))
-        landmarks_list = []
-
-        if self.verbose >= 1:
             pbar = ProgBar(len(file_list))
         for f in file_list:
+            img = read_image(filename=f, path=target_img_dir)
             if self.verbose >= 1:
                 pbar.update()
-            img = read_image(filename=f, path=target_img_dir)
 
             if self.target_width_ != img.shape[1]:
                 width_ratio = self.target_width_ / img.shape[1]
@@ -150,7 +182,8 @@ class EyepadAlign(object):
             if np.sum(landmarks) is not None:  # i.e., None == no face detected
                 landmarks_list.append(landmarks)
             else:
-                warnings.warn('No face detected in image %s. Image ignored.' % f)
+                warnings.warn('No face detected in image %s. Image ignored.'
+                              % f)
         self.target_landmarks_ = np.mean(landmarks_list, axis=0)
 
         props = self._calc_eye_properties(self.target_landmarks_)