+++ mlxtend/regressor/tests/test_stacking_cv_regression.py
@@ -1,6 +1,6 @@
 # Out-of-fold stacking regressor tests
 #
-# Sebastian Raschka 2014-2018
+# Sebastian Raschka 2014-2019
 #
 # mlxtend Machine Learning Library Extensions
 # Author: Eike Dehling <e.e.dehling@gmail.com>
@@ -124,6 +124,8 @@ def test_get_params():
               'linearregression',
               'meta-svr',
               'meta_regressor',
+              'n_jobs',
+              'pre_dispatch',
               'refit',
               'regressors',
               'ridge',